Refactor trace store manifest trace helpers
